Harden Band NemoClaw integration paths
